back/requirements.txt
cohere
python-dotenv
flask